target/risc-v: add RTT commands
The original patch to add the SEGGER RTT tooling added it only for ARM
devices. But now JLink supports RISC-V targets and we can now add the
RTT command domain to the RISC-V target. I've tested this on my RISC-V
32-bit chip and it appears to be working - I was able to communicate with
the device using the RTT.

This patch originally appeared here:
https://review.openocd.org/c/openocd/+/8234

Signed-off-by: Morozov-5F <morozov.evgeniy.95@gmail.com>
Signed-off-by: Chris Friedt <cfriedt@tenstorrent.com>